Move test to tests/plotting
