+++ pandas/tests/plotting/common.py
