+++ pandas/tests/plotting/test_boxplot_method.py
@@ -16,7 +16,7 @@ from numpy.random import randn
 
 import pandas.plotting as plotting
 
-from pandas.plotting.tests.common import (TestPlotBase, _check_plot_works)
+from pandas.tests.plotting.common import (TestPlotBase, _check_plot_works)
 
 
 """ Test cases for .boxplot method """
