+++ pandas/tests/plotting/test_converter.py
