+++ pandas/tests/plotting/test_groupby.py
@@ -8,7 +8,7 @@ import pandas.util.testing as tm
 
 import numpy as np
 
-from pandas.plotting.tests.common import TestPlotBase
+from pandas.tests.plotting.common import TestPlotBase
 
 
 @tm.mplskip
