+++ pandas/tests/plotting/test_hist_method.py
@@ -10,7 +10,7 @@ import numpy as np
 from numpy.random import randn
 
 import pandas.plotting as plotting
-from pandas.plotting.tests.common import (TestPlotBase, _check_plot_works)
+from pandas.tests.plotting.common import (TestPlotBase, _check_plot_works)
 
 
 @tm.mplskip
