+++ pandas/tests/plotting/test_misc.py
@@ -12,7 +12,7 @@ from numpy import random
 from numpy.random import randn
 
 import pandas.plotting as plotting
-from pandas.plotting.tests.common import (TestPlotBase, _check_plot_works,
+from pandas.tests.plotting.common import (TestPlotBase, _check_plot_works,
                                           _ok_for_gaussian_kde)
