+++ pandas/tests/plotting/test_series.py
@@ -17,7 +17,7 @@ import numpy as np
 from numpy.random import randn
 
 import pandas.plotting as plotting
-from pandas.plotting.tests.common import (TestPlotBase, _check_plot_works,
+from pandas.tests.plotting.common import (TestPlotBase, _check_plot_works,
                                           _skip_if_no_scipy_gaussian_kde,
                                           _ok_for_gaussian_kde)
