+++ setup.py
@@ -675,7 +675,6 @@ setup(name=DISTNAME,
                 'pandas.tests.tools',
                 'pandas.tests.types',
                 'pandas.tests.plotting',
-                'pandas.tests.test_msgpack',
                 'pandas.tools',
                 'pandas.tseries',
                 'pandas.types',
